requirements.txt
absl-py==2.2.2
accelerate==1.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alibabacloud-alimt20181012==1.1.0
alibabacloud-openplatform20191219==2.0.0
alibabacloud-tea==0.4.3
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_oss_sdk==0.1.0
alibabacloud_oss_util==0.0.6
alibabacloud_tea_console==0.0.1
alibabacloud_tea_fileform==0.0.5
alibabacloud_tea_openapi==0.3.14
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gradio==5.1.0
gradio_client==1.4.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.4
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
networkx==3.3
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.5
safetensors==0.5.3
scipy==1.15.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.2
sympy==1.13.1
tokenizers==0.20.3
tomlkit==0.12.0
tqdm==4.67.1
transformers==4.46.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
yarl==1.20.0
